Change default `hide_file_pattern ` from `NULL` to `.?*` — and uncover some problems.
First, the pattern `.`, which should work, doesn’t — because that’s the default `document_root`. In other words, that pattern hides the document root.

There are fixes for that, but they require some thinking. For example, maybe the document root (and websocket root) should always covert “.” to the absolute path to the working directory? Not sure. But there are other ways for things to go wrong. For example, what if the document root is `../example`? Or `../.`?

Anyway, second: the quick fix to that problem, using the pattern `.?*`, didn’t initially work; it looped forever. I added a patch — but left a question about a case the patch may not handle correctly.

And third: while looking at the code I immediately spotted another potential problem — and left a question about that too.

P.S. Using `.?*` works because it requires at least one char after the dot — and all dot files will have at least one char after the dot. (Of course, so will `..`…)